+++ extension/src/commands/command-controller.ts
@@ -1,9 +1,6 @@
 /* CommandController is responsible for registering possible commands */
 
-import {
-  commands
-} from 'vscode'
-
+import { commands } from 'vscode'
 import { ext } from '../main'
 import * as commandID from './command-constants'
 import { DEBUG_LOG } from '../utils/debug'
@@ -19,7 +16,6 @@ export class CommandController {
 
   // Registers a command with VS Code so it can be invoked by the user.
   #registerCommand (command: string, callback: (...args: any[]) => any): void {
-    // TODO: Do I need to subscribe the cmds?
     // DEBUG_LOG('Start Registering command ' + command)
 
     const cmd: Disposable = commands.registerCommand(command, callback)
@@ -36,11 +32,11 @@ export class CommandController {
   // those handled by the Language Server).
   #registerCommands (): void {
     DEBUG_LOG('Start Register Commands')
-    this.#registerCommand(commandID.START_EMULATOR, () => { this.#startEmulator() })
-    this.#registerCommand(commandID.STOP_EMULATOR, () => { this.#stopEmulator() })
-    this.#registerCommand(commandID.RESTART_SERVER, () => { this.#restartServer() })
-    this.#registerCommand(commandID.CREATE_ACCOUNT, () => { this.#createAccount() })
-    this.#registerCommand(commandID.SWITCH_ACCOUNT, () => { this.#switchActiveAccount() })
+    this.#registerCommand(commandID.START_EMULATOR, this.#startEmulator)
+    this.#registerCommand(commandID.STOP_EMULATOR, this.#stopEmulator)
+    this.#registerCommand(commandID.RESTART_SERVER, this.#restartServer)
+    this.#registerCommand(commandID.CREATE_ACCOUNT, this.#createAccount)
+    this.#registerCommand(commandID.SWITCH_ACCOUNT, this.#switchActiveAccount)
     DEBUG_LOG('Register Commands Done')
   }
