+++ extension/src/emulator/local/config.ts
@@ -1,5 +1,5 @@
 /* Handle flow.json config file */
-import { window, workspace } from 'vscode'
+import { window, workspace, commands } from 'vscode'
 import * as util from 'util'
 import * as cp from 'child_process'
 import { FILE_PATH_EMPTY } from '../../utils/utils'
@@ -13,6 +13,7 @@ let configPath: string | undefined
 export async function getConfigPath (): Promise<string> {
   if (configPath === undefined) {
     configPath = await retrieveConfigPath()
+    handleConfigChanges()
   }
   return configPath
 }
@@ -60,91 +61,32 @@ async function readLocalConfig (): Promise<string> {
   return file[0].path
 }
 
-// TODO: HandleConfigChanges()?
-
-/*
-export class Config {
-  // Full path to flow.json file
-  static configPath: string
-
-  private constructor () {}
-
-  static async getConfigPath (): Promise<string> {
-    if (!Config.configPath) {
-      Config.configPath = await Config.#retrieveConfigPath()
+// Called when configuration is changed
+function handleConfigChanges (): void {
+  workspace.onDidChangeConfiguration((e) => {
+    // TODO: do something smarter for account/emulator config (re-send to server)
+    const promptRestartKeys = [
+      'languageServerPath',
+      'accountKey',
+      'accountAddress',
+      'emulatorAddress'
+    ]
+    const shouldPromptRestart = promptRestartKeys.some((key) =>
+      e.affectsConfiguration(`cadence.${key}`)
+    )
+    if (shouldPromptRestart) {
+      window
+        .showInformationMessage(
+          'Server launch configuration change detected. Reload the window for changes to take effect',
+          'Reload Window',
+          'Not now'
+        )
+        .then((choice) => {
+          if (choice === 'Reload Window') {
+            commands.executeCommand('workbench.action.reloadWindow')
+              .then(() => {}, () => {})
+          }
+        }, () => {})
     }
-    return Config.configPath
-  }
-
-  static async #retrieveConfigPath (): Promise<string> {
-    // Configuration File
-    let configPath = await Config.#readLocalConfig()
-    if (configPath === FILE_PATH_EMPTY) {
-      if (!Config.#promptInitializeConfig()) { throw Error('No valid config path') }
-      configPath = await Config.#readLocalConfig()
-    }
-    return configPath
-  }
-
-  static async #promptInitializeConfig (): Promise<boolean> {
-    let rootPath: string | undefined
-    if ((workspace.workspaceFolders != null) && (workspace.workspaceFolders.length > 0)) {
-      rootPath = workspace.workspaceFolders[0].uri.fsPath
-    } else {
-      rootPath = workspace.rootPath // ref: deprecated
-    }
-    if (rootPath === undefined) {
-      return false
-    }
-
-    const continueMessage = 'Continue'
-    const selection = await window.showInformationMessage('Missing Flow CLI configuration. Create a new one?', continueMessage)
-    if (selection !== continueMessage) {
-      return false
-    }
-
-    await exec('flow init', { cwd: rootPath })
-
-    return true
-  }
-
-  static async #readLocalConfig (): Promise<string> {
-    const file = await workspace.findFiles('flow.json')
-    if (file.length !== 1) {
-      return FILE_PATH_EMPTY
-    }
-    return file[0].path
-  }
-
-  // TODO: What needs to call this?
-  // Adds an event handler that prompts the user to reload whenever the config changes
-  static #handleConfigChanges (): void {
-    workspace.onDidChangeConfiguration((e) => {
-      // TODO: do something smarter for account/emulator config (re-send to server)
-      const promptRestartKeys = [
-        'languageServerPath',
-        'accountKey',
-        'accountAddress',
-        'emulatorAddress'
-      ]
-      const shouldPromptRestart = promptRestartKeys.some((key) =>
-        e.affectsConfiguration(`cadence.${key}`)
-      )
-      if (shouldPromptRestart) {
-        window
-          .showInformationMessage(
-            'Server launch configuration change detected. Reload the window for changes to take effect',
-            'Reload Window',
-            'Not now'
-          )
-          .then((choice) => {
-            if (choice === 'Reload Window') {
-              commands.executeCommand('workbench.action.reloadWindow')
-                .then(() => {}, () => {})
-            }
-          }, () => {})
-      }
-    })
-  }
+  })
 }
-*/
