+++ extension/test/e2e/send-script.test.ts
