+++ extension/test/e2e/switch-accounts.test.ts
