+++ extension/test/helpers.ts
